custom-estimator/requirements.txt
scikit-learn == 0.19.2
scipy == 1.1.0
keras == 2.2.2
pillow == 5.2.0
pandas == 0.23.4
MultiColProcessor == 1.0.15
tensorflow-gpu == 1.3.0
matplotlib == 2.2.3
opencv-python == 3.4.2.17
np-utils == 0.5.5.2





